init-ceph.in: skip ceph-disk if it is not present
With the ceph-mon vs ceph-osd split packaging, users are expected to
have the ceph-mon package installed and not ceph-osd (and vice versa).
However, the init script (/etc/init.d/ceph) has a call to `ceph-disk`,
which may not be present on the machine.

Given that our packaging is not yet split upstream, this bug does not
manifest itself currently, because both ceph-mon and ceph-disk are
currently in the same "ceph" package. Once we split the packaging,
though, this will become an issue.

http://tracker.ceph.com/issues/10587 Refs: #10587

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>
(cherry picked from commit 7dd9af0a21e0ab0ae397faf901231ad0a9688ab9)

Resolves: rhbz#1225183