man: document listwatchers cmd in "rados" manpage
This is the downstream patch to resolve rhbz#1293555, "'rados' man page
doesn't mention the 'listwatchers' switch".

It is different than Kefu's patch upstream (identical title), because
that patches the rST source, whereas I need to patch the nroff code that
upstream ships in the release tarball.

When upstream ships an updated man/rados.8 file, we can drop this patch.