rgw: use bucket_info.bucket_id instead of marker where needed
We used to use these interchangeably, but they actually have diffent meaning.
The marker is the prefix we assign to the objects in this bucket, whereas
the bucket_id is the bucket instance's id. These used to hold the same
value, but with bucket resharding it's not going to be true anymore.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>
(cherry picked from commit d41cbb44c55cf2dca5ab25d5a1ec76e30db26150)

Resolves: rhbz#1378995