rgw: bucket link: "bucket move"; handle bucket names too.
This is the remainder of bucket move function.  Implement bucket "rename"
- use string passed in via '--bucket-new-name' to optinally override
the resultant bucket name when doing bucket link.  This is basically
a slight generalization of changing the tenant of a bucket; the same
operations are required for either.

Further refinements included here are minor improvements to associated
error messages.

This does not change any "tail placement" information in any objects
contained in the bucket.  The bucket name is encoded there, along with
the bucket id, but neither appears to be used, and the existing reshard
logic which changes bucket ids also does not appear to alter that.

Fixes: http://tracker.ceph.com/issues/35885
Resolves: rhbz#1763219

Signed-off-by: Marcus Watts <mwatts@redhat.com>
(cherry picked from commit b908e45c024d96ff973d9eddd8a96bcddc5583bf)