librbd: use custom allocator for aligned boost::lockfree::queue
If tcmalloc is in-use as the allocator and its version is less than 2.6.2,
it might be missing support for 'aligned_alloc'. This can result in the
glibc version of 'aligned_alloc' being used to allocate memory that is
then freed by tcmalloc -- resulting in a crash.

Fixes: http://tracker.ceph.com/issues/39703
Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 7f6685383dfdab1b685c9c4df609feb7fdc97a91)

Conflicts:
	CMakeLists.txt: trivial resolution
	src/librbd/ImageCtx.h: dropped lock-free changes but kept allocator

Resolves: rhbz#1860073