elector: support disallowing specific monitor ranks from becoming leaders
If they are disallowed, they always defer to others and are never
deferred to themselves. You can use this if you have a high-latency
monitor you want as a peon but not a leader, for instance.
Currently there is no way to configure it within the wider ecosystem, though.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 78739e616178098c639957f5d7482affb18100f1)

Resolves: rhbz#1800382