rgw: make rollback refcount tag match
When the refcount does a "get", the tag ends with a null character,
but during a rollback, when the tag is "put" it does not end with a
null character, so the same tag is not being manipulated and the
refcounts are therefore off, preventing future deletion of an rgw
object. This adds the null character to the "put" operation.

This change was originally submitted by lltlo <120680451@qq.com> but
wasn't signed-off.

Resolves: rhbz#1762926

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit a0fee94982d89f0210e1c526605831105cd90ef4)
(cherry picked from commit e8216d06d39bd854180b1c060ca794d21095e476)