rgw: define rgw_bucket::full_equal(...) method
The  rgw_bucket::operator== method considers bucket name and id,
but not tenant--but operator!= considers all three.

It seems clear that operator== should be full and have symmetry
with !=, but to avoid unexpected side effects, define a new
full_equal(...) comparison that considers name, id, and tenant,
and use it in the bucket-link same-bucket-and-tenant test.

Resolves: rhbz#1763219

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>