rgw: add radosgw-admin subcommand "bucket radoslist"
This subcommand takes a bucket as a command-line argument and walks
the bucket listing and that manifest (if it exists) for each object in
the listing to generate the rados objects that represent the objects
in the bucket.

This has been tested with large objects (heads + tails), s3 multipart
objects, versioned buckets, resharded buckets, and buckets with
objects copied into them to cover a variety of variations.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#1752130