elector: ignore election proposals from out-of-quorum spammers
Remember the last leader we *voted* for in an election. Then, if we get
a proposal from an out-of-quorum peer, and we would vote for the same
leader as we did in the last successful election, and we aren't the
leader...just drop it.

The leader can call a new election if the out-of-quorum peer can reach
it, and we will continue pinging so scores may change in the future,
but we don't want to thrash elections.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit a540bed4fa8b6e82f915e0b6e3842b1163a67ab4)

Resolves: rhbz#1800382