elector: maintain a stable score set while running an election
Previously we could update our scores during an election, and
then try to change who we voted for if we bumped the epoch. (Or,
mostly, just assert because thankfully we had checks for these
sorts of logic failures!) That was bad. Instead, whenever we start()
a new election, make a stable copy of the ConnectionTracker we use
for the duration, and clean it up when the election ends.

If the state materially changes during the election period, that's
okay -- another election will be triggered by somebody timing out
or trying to join, and we'll use the updated scores for that one!

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 020d2e8cddbd1b575997cf9c1fcd2e6dd6f68341)

Resolves: rhbz#1800382