rgw: allow reshard log entries for non-existent buckets to be cancelled
The radosgw-admin tool allows admins to add buckets to the reshard log
and to cancel buckets from the reshard log. Both operations check for
the existence of the bucket before proceeding and fail for nonexistent
buckets.

It's possible, however, for an admin to add a bucket to the reshard
log and then, before the bucket is resharded, for a user to delete the
bucket. This leaves the entry in the reshard log.

Prior to this commit an attempt to use radosgw-admin to cancel the
reshard log entry would fail. With this commit it will still fail
*but* notify the user they can use the --yes-i-really-mean-it
command-line option to do it nonetheless. And if the user includes
that option, it will succeed.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 9f893a7398716127711a0baaa937070a495a7a56)

Resolves: rhbz#1785080