rgw: mismatched indices in parallel vectors in RGWRados::cls_bucket_list_ordered
cls_bucket_list_ordered maintains three parallel vectors to process
entries from each bucket index shard. The vectors had been constructed
incorrectly to add a series of unneeded entries at the beginning. A
recent change removed the unneeded entries at the start of one of the
vectors but not the other two. That allowed an index determined from
one vector to cause an access beyond the end of a different vector.

This fix removes the unneeded entries from the other two vectors so
the valid indices are the same across the three vectors.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
Resolves: rhbz#1766448