rgw: auto-clean reshard queue entries for non-existent buckets
It is possible for a bucket to be added to the reshard queue and then
to be removed before its entry in the reshard queue is processed. When
this is now encountered, processing of the reshard queue errors out.

This fix recognizes when the reshard queue entry refers to a
non-existent bucket and remove the entry from the reshard queue,
allowing processing of the queue to continue.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 304ad8803292e92911e86b84c39a6c2ff0d0121b)

Resolves: rhbz#1833497