rgw: add tooling to support rgw-orphan-list
This adds tooling to support the listing of orphans in a pool.

This adds a binary tool ceph-diff-sorted that will perform a diff on
two files sorted with "LANG=C".

This also adds a bash script named rgw-orphan-list that will produce a
list in a local file of what appear to be rgw orphans.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#1815211