librbd: relax snapshot mirror promotion state determination
If no mirroring-related snapshots exist on the image, it cannot
be determined if the image is primary/non-primary. In the case of
rbd-mirror, this will be the case after a new mirrored image is
created locally but before it can be synced.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit 21c51c232ba7a06f8dd78be3af7331bca8da7c6b)

Resolves: rhbz#1860073