rbd-mirror: initial snapshot-based mirroring replayer logic
This initial implementation will scan the remote image for new mirror
snapshots. If a new primary snapshot it detected, it will start syncing
it. If the remote was demoted, it will notify the image replayer to
shut down. If an in-progress sync was previously interrupted, it will
attempt to restart it.

Signed-off-by: Jason Dillaman <dillaman@redhat.com>
(cherry picked from commit e4cb7e4d78bcfc3cc96eea5313c101044013cb7b)

Resolves: rhbz#1860073