rgw: Fix for correctly listing all elements of queue.
The gc list operations lists 1000 elements at a time,
and if 1000 elements are returned from a queue, then
the next iteration also lists the same queue to see
if there are some more elements. This fix takes care
of cases when there are more than 1000 elements in the
queue and the number of elements is a multiple of 1000.
The truncated flag of queue list is also taken into account
to determine whether we should continue processing the
same queue or move on to the next gc object.

Resolves: rhbz#1738334
Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>