mon: add crush_loc to mon_info_t
Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 0af65d60c53901e68135c48d87995bf963ac9b30)

Conflicts:
	src/mon/MonMap.cc

We have to adjust the mon_info encoding strategy a bit:
the crush_location was written after a new struct version 4
already existed, adding "weight" to the mon_info_t structure. So
we here fill it in with the default value, assert that we only
see the default value when decoding, and then ignore it.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
Resolves: rhbz#1800382