rgw: don't write to error repo for async notification failures
async notifications are just hints, and don't imply an obligation to
sync the bucket shard. if we fail to sync, don't write it to the error
repo for retry. we'll see the change later when processing the datalog

Signed-off-by: Casey Bodley <cbodley@redhat.com>
(cherry picked from commit af2114a69b85b0cfdeb69c4c470abbec3eb50758)
Resolves: rhbz#1964097
Signed-off-by: Adam C. Emerson <aemerson@redhat.com>