rgw: find oldest period and update RGWMetadataLogHistory()
when a period lookup for oldest_realm_epoch returns an ENOENT,
find the oldest one and update RGWMetadataLogHistory. This is to avoid
an
empty cursor being passed in to ceph_assert() in
PurgePeriodLogsCR::operate()
in case of incomplete period history.

Signed-off-by: Shilpa Jagannath smanjara@redhat.com

Fixes: https://tracker.ceph.com/issues/40341

Resolves: rhbz#1552567

(backport of upstream commit 0ea65d81d4d5431ca22340149dbd714e395ecdc3
to downstream 3.3z)

Signed-off-by: Mark Kogan <mkogan@redhat.com>