rgw: attempt retries after write race for bucket reshard bucketinfo update
During bucket resharding, if there is a race and write to upate the
BucketInfo object with the status of resharding gets cancelled due to
that race, resharding fails. This change will retry the update up to
five times, so that resharding can continue. Furthermore, with each
retry there's exponential back-off.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#1835777