rgw: orphan-list timestamp fix
When creating intermediate and output files, the rgw-orphan-list
script uses a timestamp using the `date` command. The hour was
inserted with "%k" but that padds with a space rather than a zero. So
that's changed to "%H".

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit b1daf30d61989d0813ea73ac19f36b4de6abf13d)

Resolves: rhbz#1855380