rgw/sts: adding code for federated user as owner in case of STS.
A new user under the namespace 'oidc' is created for every federated
user in case of AssumeRoleWithWebIdentity.

In case of AssumeRole, the user that needs cross account access becomes
the owner.

In both cases buckets are created in the tenant that the role belongs to.

Fixes:rhbz#1976883

Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>
(cherry picked from commit 94f1b974c54868807d58ed39da800a20d0d5df47)