mgr: Add python-enum34 dependency to package for older distributions
Subvolumes v2 introduces Enum usage, that is not part of python<=3.4 and is
provided as an extra package python-enum34 till python 2.4 version.

Added this as an explicit dependency to packaging, to ensure it is installed
in required distributions.

python-enum34 is also used by mgr-dashboard
(src/pybind/mgr/dashboard/plugins/feature_toggles.py) but the dependency is
not called out explicitly and is satisfied by, python-pyopenssl, which
depends on python-cryptography, which depends on python-enum34.

Platform availability notes:
- centos7 comes with python-enum34
- centos8 does not require this, as python versions are higher, and Enum is part
of the language itself
- openSUSE 15.1/2 comes with python-enum34
- Ubuntu Xenial/Bionic comes with python-enum34

Signed-off-by: Shyamsundar Ranganathan <srangana@redhat.com>
(cherry picked from commit fdfae1c00c0059a03389bdcfafa6b0ee8d95ad55)

Resolves: rhbz#1848494