rgw: delimited ordered bucket list doesn't skip over unneeded entries
This sets the skip_after_delim variable whenever an entry is
(re-)added to the common_prefixes collection. There is existing logic
to use skip_after_delim to advance the marker for the next call to
RGWRados::cls_bucket_list_ordered to skip over unneeded entries.

Further optimizations could be made, but this resolves the proximate
issue.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#1874645