rgw: during GC defer, prevent new GC enqueue
With the new queue-based GC code, when a GC defer operation is
performed, it adds an "urgent" record to prevent GC from occuring,
whether there's a GC entry or not (it's not checked).

But either way the code *also* adds a new GC entry to the queue to
cause GC to occur at a later time. This would be incorrect if there is
no GC entry to begin with. This will cause GC to delete tail objects
when there has been no user-initiated delete. In other words a READ
operation can result in a permenent DELETE of portions of large
objects.

This is a temporary fix for this bug. It marks the code in error and
prevents GC defer operations from taking place at all as a temporary
measure.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>

Resolves: rhbz#1892644