mon: fix some erroneous safety asserts in OSDMonitor::try_enable_stretch_mode
This function asserts it is not committing in error cases, because it is supposed
to be run first in a trial mode, and then in a commit mode. But I wanted the
asserts to be more informative than that when they are printed, so I included
the failure condition as an OR clause in the assert.

Unfortunately, a few of them were either written wrong or became wrong due to
code movement. Fix these so they are valid and convey the desired information.

Reported-by: Ronen Friedman <rfriedma@redhat.com>

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit 208957e6d99f1dfe83e5601f6db80a90a96b62cd)

Resolves: rhbz#1800382