rgw/sts: adding code for aws:RequestTags as part of session tags implementation.
Session Tags can be passed in the web token in
AssumeRoleWithWebIdentity call by configuring
them in the IDP. These tags can be used as Conditions
in the trust policy of a role as aws:RequestTag, based
on which a federated user is allowed to assume a role.
The trust policy should have a statement for 'sts:TagSession'
alongwith 'sts:AssumeRoleWithWebIdentity' in case principal
tags are passed in the web token.

fixes:rhbz#1992145

Signed-off-by: Pritha Srivastava <prsrivas@redhat.com>
(cherry picked from commit ce632c96f7204f6189aeafedbebce880f76b153b)