osd: PeeringState: fix calc_replicated_acting_stretch() syntax/logic
There was a major error here! get_ancestor() was type-deduced to return
a bucket_candidates_t -- a *copy* of what was in the map, not the reference
to it we wanted to actually amend!

Fix this by returning a pointer instead. There's a way to coerce things
to return a reference instead but the syntax seems clumsier to me
and I'm not familiar with it anyway -- this works just fine.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>
(cherry picked from commit d09253081f7703fcb5a1bebce8840b489a3fc3ab)

Conflicts:
	src/osd/PeeringState.cc

Signed-off-by: Greg Farnum <gfarnum@redhat.com>

Resolves: rhbz#1937088