mon: monmap: fix forward- and backward-compatibility of mon_info_t encoding
The initial commit[1] backporting mon locations into mon_info_t needed to account
for the update which also added monitor weights. It made an error: the default
weight was set to 10 in the commit which *added* weights, but was immediately
changed to 0 before any releases occurred. And then it asserted the decoded
value matched the default, because the nautilus backport is just throwing this
data away. But this code is run on clients as well as servers!

So, toss out the assert completely, and encode the weight as 0.

Secondly, I fixed a similar issue where I'd set min_compat_v too high in
"mon: Do not increase compatv when using monitor location or stretch mode",
but I set the min_compat_v to 3. This is incorrect -- up until I started
mucking around, the min_compat_v was 1, so that old clients can continue
decoding things. And that is still fine here.

[1]: "mon: add crush_loc to mon_info_t"

Signed-off-by: Greg Farnum <gfarnum@redhat.com>

Resolves: rhbz#1936210