rgwlc: add a wraparound to continued shard processing
If the full set of buckets for a given lc shard couldn't be
processed in the prior cycle, processing will start with a
non-empty marker.  Note the initial marker position, then
when the end of shard is reached, allow processing to wrap
around to the logical beginning of the shard and proceeding
through the initial marker.

Resolves: rhbz#1887661

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit 4e15e76aca4dc198d7f1b6dde7bb5e068dbb961e)