auth/cephx: cap ticket validity by expiration of "next" key
If auth_mon_ticket_ttl is increased by several times as done in
commit 522a52e6c258 ("auth/cephx: rotate auth tickets less often"),
active clients eventually get stuck because the monitor sends out an
auth ticket with a bogus validity.  The ticket is secured with the
"current" secret that is scheduled to expire according to the old TTL,
but the validity of the ticket is set to the new TTL.  As a result,
the client simply doesn't attempt to renew, letting the secrets rotate
potentially more than once.  When that happens, the client first hits
auth authorizer errors as it tries to renew service tickets and when
it finally gets to renewing the auth ticket, it hits the insecure
global_id reclaim wall.

Cap TTL by expiration of "next" key -- the "current" key may be
milliseconds away from expiration and still be used, legitimately.
Do it in KeyServerData alongside key rotation code and propagate the
capped TTL to the upper layer.

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>
(cherry picked from commit 370c9b13970d47a55b1b20ef983c6f01236c9565)

Conflicts:
	src/auth/cephx/CephxKeyServer.cc [ commits ef3c42cd6481 ("auth:
	  EACCES, not EPERM") and 83fbd116e208 ("auth: switch from Mutex
	  to ceph::mutex") not in luminous ]
	src/auth/cephx/CephxServiceHandler.cc [ commits 5df1a2ee9e8c
	  ("auth: remove auid member from EntityAuth") and 1adf313ef94a
	  ("auth/cephx: share all tickets and connection_secret in
	  initial reply") not in luminous ]

Resolves: rhbz#1939092